Fix extensions for OSX
Improper extension was being returned for OSX/Darwin,
causing an error if the LD_LIBRARY_PATH was not set.

Previously any non-windows system would attempt to use
'.so' as the library suffix, causing it to fail on Darwin,
which should be looking for '.dylib'. Fix was just to rely
on the FFI gem to detect and return the proper value and not
let this gem decide.