Add reference counter for LIBUSB::Context usage
The reference counter is increased on every object that is bound to a libusb context.
If the context is freed before any object that is bound to a context, then the call to `libusb_exit` is postponed to the last unreferening object.
That avoids segfaults caused by logging through the context when other objects are freed.

This patch changes several objects from FFI::ManagedStruct to FFI::Struct with manual definition of a finalizer.
This is because we can not easily pass a ctx pointer to the ManagedStruct release method.

The Call::Transfer is independent from the context and doesn't need the reference counting.

Fixes #46