Deregister callbacks when Context is GC'ed and on eventmachine_unregister
Also add the ability to unregister the callback through the API and document this.

Unregistering callbacks in eventmachine part was a missing cleanup step.
And because the callbacks are typically GC'ed before the LIBUSB::Context, they are no longer valid when libusb_exit is called, leading to segfauls.

Fixing both parts makes sure, that ressources are properly freed and even if not, they don't cause a segfault.