Disable linux fat binary builds
There is the issue, that linux binary builds depend on libudev to be installed on the target system and not all distributions provide it.
The solution would be to static link the fat binary libusb to libudev, but libudev is part of systemd and that needs meson and ninja in very recent versions, making it hard to buill in the RCD environment.
So for now linux binary gems are not distributed.